requirements.txt
django
djangorestframework
pandas
numpy
scikit-learn
joblib
django-cors-headers
